dockerized

+++ build.sh
@@ -2,25 +2,36 @@
 
 CMAKE_BIN=cmake
 
+TARGET_PLATFORM=unix
+
 BUILD_TYPE=Debug
+SANETIZE=none
 
-PROJECT_DIR=`pwd`
-BUILD_DIR=build/${BUILD_TYPE}
+test "x${PET_CONSOLE_TARGET_PLATFORM}" = 'xemscripten' && TARGET_PLATFORM=${PET_CONSOLE_TARGET_PLATFORM}
+test "x${PET_CONSOLE_BUILD_TYPE}" != 'x' && BUILD_TYPE=${PET_CONSOLE_BUILD_TYPE}
+test "x${PET_CONSOLE_SANETIZE}" != 'x' && SANETIZE=${PET_CONSOLE_SANETIZE}
 
-if [[ "xrebuild" == "x$1" ]]; then
-	rm -rf ${BUILD_DIR}
-fi
-mkdir -p ${BUILD_DIR}
+echo "target platform [${TARGET_PLATFORM}]"
+echo "build type [${BUILD_TYPE}]"
+echo "sanetize [${SANETIZE}]"
+
+APP_NAME=pet_console
+
+PROJECT_DIR=$PWD
+BUILD_DIR=${PROJECT_DIR}/.build/${TARGET_PLATFORM}/${BUILD_TYPE}
+BIN_DIR=${BUILD_DIR}/bin
+
+test "x${1}" = 'xrebuild' && rm -rf ${BUILD_DIR}
+${CMAKE_BIN} -E make_directory ${BUILD_DIR}
+
+CMAKE_FLAGS="-DCMAKE_INSTALL_PREFIX=${BUILD_DIR}"
+CMAKE_FLAGS="${CMAKE_FLAGS} -DCMAKE_BUILD_TYPE=${BUILD_TYPE}"
+
+. ./scripts/building/platform/${TARGET_PLATFORM}.sh
+
 cd ${BUILD_DIR}
-
-${CMAKE_BIN} -DCMAKE_BUILD_TYPE=${BUILD_TYPE} ${PROJECT_DIR}
-${CMAKE_BIN} --build . -- -j4
-
-mkdir -p bin
-cp -v src/app/app bin
-
+build_platform_create_project
 cd ${PROJECT_DIR}
 
-echo
-echo run:
-echo ./${BUILD_DIR}/bin/app
+${CMAKE_BIN} --build ${BUILD_DIR} --target install -- -j 8
+build_platform_create_bin
